+++ src/main/java/cn/widealpha/train/domain/Station.java
@@ -2,18 +2,18 @@ package cn.widealpha.train.domain;
 
 public class Station {
 
-  private long stationId;
+  private Integer stationId;
   private String name;
   private String telecode;
   private String en;
   private String abbr;
 
 
-  public long getStationId() {
+  public Integer getStationId() {
     return stationId;
   }
 
-  public void setStationId(long stationId) {
+  public void setStationId(Integer stationId) {
     this.stationId = stationId;
   }
